move wasm into another crate
