+++ diagram-editor/Cargo.toml
@@ -41,14 +41,11 @@ indexmap = { version = "2.10.0", optional = true, features = ["serde"] }
 mime_guess = { version = "2.0.5" }
 schemars = { version = "0.9", optional = true }
 serde = "1.0.219"
-serde-wasm-bindgen = { version = "0.6.5", optional = true }
 serde_json = "1.0.140"
 tar = { version = "0.4.44", optional = true }
 thiserror = "2.0.16"
 tokio = { version = "1.45.1", features = ["macros", "time"] }
 tracing = "0.1.41"
-wasm-bindgen = { version = "0.2.100", features = ["serde"], optional = true }
-wasm-bindgen-futures = { version = "0.4.50", optional = true }
 
 [build-dependencies]
 flate2 = { version = "1.1.1", optional = true }
@@ -66,12 +63,6 @@ debug = ["router", "axum/ws"]
 frontend = ["router", "dep:flate2", "dep:tar"]
 router = []
 json_schema = ["dep:schemars", "dep:indexmap"]
-wasm = [
-  "dep:wasm-bindgen-futures",
-  "dep:serde-wasm-bindgen",
-  "dep:wasm-bindgen",
-  "bevy_impulse/single_threaded_async",
-]
 
 [[bin]]
 name = "print_schema"
