+++ diagram-editor/server/lib.rs
@@ -1,5 +1,3 @@
-use crate::api::ApiOptions;
-
 #[cfg(feature = "router")]
 pub use router::*;
 #[cfg(feature = "router")]
@@ -9,9 +7,3 @@ pub mod api;
 
 #[cfg(feature = "frontend")]
 mod frontend;
-
-#[derive(Default)]
-#[non_exhaustive]
-pub struct ServerOptions {
-    pub api: ApiOptions,
-}
