+++ diagram-editor/wasm/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.0.1"
 edition = "2021"
 authors = ["Teo Koon Peng <koonpeng@intrinsic.ai>"]
 license = "Apache-2.0"
-description = "Frontend for bevy_impulse diagrams"
+description = "Webassembly backend for bevy_impulse diagram editor"
 readme = "README.md"
 repository = "https://github.com/open-rmf/bevy_impulse"
 keywords = [
@@ -13,8 +13,9 @@ keywords = [
   "behavior",
   "agent",
   "bevy",
-  "frontend",
   "diagram",
+  "wasm",
+  "webassembly",
 ]
 categories = [
   "science::robotics",
@@ -27,8 +28,17 @@ categories = [
 crate-type = ["cdylib"]
 
 [dependencies]
-axum = { version = "0.8.4" }
-# bevy_impulse_diagram_editor = { version = "0.0.1", path = "..", default-features = false }
+axum = { version = "0.8.4", features = ["json"], default-features = false }
+bevy_app = "0.12.1"
+bevy_impulse = { version = "0.0.2", path = "../..", features = [
+  "single_threaded_async",
+] }
+bevy_impulse_diagram_editor = { version = "0.0.1", path = "..", default-features = false }
+mime_guess = { version = "2.0.5" }
+serde_json = "1.0.140"
+serde-wasm-bindgen = "0.6.5"
+wasm-bindgen = "0.2.100"
+wasm-bindgen-futures = "0.4.50"
 
 [build-dependencies]
