+++ diagram-editor/wasm/src/lib.rs
@@ -1,1 +1,6 @@
+mod errors;
+mod executor;
+mod globals;
 
+pub use executor::*;
+pub use globals::*;
